Remove "Our Next Meeting" feature
Meetup removed the API Key Signature feature in favor of Oauth 2, which
doesn't really work with the way the this site has been designed
(everything is client-side code). I also don't want to have individual
users have to be logged in to Meetup to see the event information.

Until I figure out another way to handle this, or just replace the
section with something simpler, I'm just going to remove the section
entirely.